fiddling with sontek adp code
This produces a *lot* of debugging output.  I was trying to guess the
checksum convention, but I failed to make any significant progress on
that.  Still, I left the debugging enabled, because I ought to be the
only person using this branch, and I will likely need this debugging
information when (or if) I get Sontek docs on the argonaut adp format.