some sontek argonaut tests (failing)
It seems that we have 161 byte chunks, but there's no such thing in the
docs. (Although two things sum to 160 or so bytes ...)

I think the docs are not for this data type.

NOTE: this code likely breaks old sontek ADP or ADV reading, so I'll
want to check that with some files at work.  What I'm doing is to
isolate the argonaut code because doing a lacework of conditionals is
more confusing than just duplicating and then altering blocks of code.