argonaut: prit some more bytes
also exports BUF and OFF.  Note that
```R
> BUF[OFF+seq(0:10)]
 [1] b5 c2 00 00 00 00 00 00 e8 03 00
```
so maybe 0xB5 is a header byte. Searching for recurrances of that value,
the separation 161 seems to be quite common. However, the byte *after*
that seems almost random, so this does not look like 0xB5 is the first
byte of a two-byte sequence. Also, if the usual nortek pattern is
followed, then the byte (or 2 bytes) after that would indicate a number
of bytes in the chunk, but in this case that number would be zero.

So, I'm no closer to guessing a format.